Large refactor of code structure.
Most classes now declared in header files and defined in source files.

Notably, the MacroForest class and its nested classes have been made into separate classes.

Some structs (e.g., TransformedDefinition, MacroArgument) have been turned into classes.

Some things left to do:
- Check for memory leaks
- Separate remaining classes into header and source files
- Remove the use of the friend keyword from some classes, and instead favor getter methods
