+++ implementation/src/CMakeLists.txt
@@ -34,11 +34,22 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O1")
 add_library(Cpp2C SHARED
   Cpp2C.cc
   CollectDeclNamesVisitor.hh
+  ExpansionUtils.cc
   ExpansionUtils.hh
+  ForestCollector.cc
+  ForestCollector.hh
+  MacroArgument.cc
+  MacroArgument.hh
+  MacroExpansionNode.cc
+  MacroExpansionNode.hh
+  MacroForest.cc
   MacroForest.hh
   MacroNameCollector.hh
   Matchers.hh
   NodeCollector.hh
+  SourceRangeCollection.cc
+  SourceRangeCollection.hh
+  TransformedDefinition.cc
   TransformedDefinition.hh
   )
