+++ implementation/src/MacroArgument.cc
@@ -1,0 +1,8 @@
+#include "MacroArgument.hh"
+
+MacroArgument::MacroArgument(const string &Name) : Name(Name) {}
+
+set<const Stmt *> MacroArgument::getStmts()
+{
+    return Stmts;
+}
